services/auth_service/requirements.txt
Django
djangorestframework
psycopg2-binary
PyJWT
django-cors-headers
python-decouple
resend
dnspython
 